// src/components/profile/types.ts
export interface UserInfo {
  maNguoiDung: number;
  tenDangNhap: string;
  email: string;
  hoTen: string;
  soDienThoai: string;
  ngaySinh: string;
  gioiTinh: string;
  diaChi: string;
  tienSuBenh?: string;
  diUng?: string;
  maBenhNhan: number;
}

// Appointment related types
export interface Appointment {
  maLichHen: number;
  maBenhNhan: number;
  tenBenhNhan: string;
  soDienThoaiBenhNhan: string;
  maBacSi: number;
  tenBacSi: string;
  maDichVu: number;
  tenDichVu: string;
  ngayHen: string;
  gioBatDau: string;
  gioKetThuc: string;
  maTrangThai: number;
  tenTrangThai: string;
  ghiChu?: string;
  lydo?: string | null;
  thoiGian: number;
}

export interface Doctor {
  maBacSi: number;
  hoTen: string;
}

export interface TimeSlot {
  gioBatDau: string;
  gioKetThuc: string;
  daDat: boolean;
  trangThai: string;
}

export interface Service {
  MaDv: string;
  TenDv: string;
}